refactor healing
